BEO-5448 - fixes and improvements

+++ WebAPI/Reflection/DataModel.cs
@@ -8439,7 +8439,7 @@ namespace WebAPI.Reflection
                             
                         case "update":
                             RolesManager.ValidateActionPermitted("segmentationType", "update", false);
-                            return SegmentationTypeController.Update((int) methodParams[0], (KalturaSegmentationType) methodParams[1]);
+                            return SegmentationTypeController.Update((long) methodParams[0], (KalturaSegmentationType) methodParams[1]);
                             
                     }
                     break;
@@ -15354,7 +15354,7 @@ namespace WebAPI.Reflection
                         case "update":
                             ret.Add("segmentationTypeId", new MethodParam(){
                                 NewName = newParamName,
-                                Type = typeof(int),
+                                Type = typeof(long),
                             });
                             ret.Add("segmentationType", new MethodParam(){
                                 NewName = newParamName,
